4.21*: bloglist expansion, step9

+++ part4/blog/controllers/blogs.js
@@ -11,15 +11,12 @@ blogRouter.get("/", async (request, response) => {
 
 blogRouter.post("/", async (request, response) => {
   const token = request.token;
-  console.log(token);
 
   const decodedToken = jwt.verify(token, process.env.SECRET);
   if (!token || !decodedToken.id) {
     return response.status(401).json({ error: "token missing or invalid" });
   }
   const user = await User.findById(decodedToken.id);
-
-  logger.info(user);
 
   const blogReq = request.body;
   const blog = new Blog({
@@ -31,22 +28,41 @@ blogRouter.post("/", async (request, response) => {
   });
 
   const result = await blog.save();
-  console.log(result);
   user.blogs = user.blogs.concat(result._id);
   await user.save();
   response.status(200).json(result);
 });
 
 blogRouter.delete("/:id", async (req, resp) => {
-  const id = req.params.id;
-  await Blog.findByIdAndDelete(id);
-  resp.sendStatus(204);
+  const token = req.token;
+
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return resp.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(req.params.id);
+
+  if (!blog) {
+    resp.sendStatus(404);
+  } else if (user.id.toString() === blog.user.toString()) {
+    const id = req.params.id;
+    await Blog.findByIdAndDelete(id);
+    resp.sendStatus(204);
+  } else {
+    resp.sendStatus(401);
+  }
 });
 
 blogRouter.put("/:id", async (req, resp) => {
   const id = req.params.id;
   result = await Blog.findByIdAndUpdate(id, req.body);
-  resp.send(200).json(result);
+  if (result) {
+    resp.status(200).json(result);
+  } else {
+    resp.sendStatus(404);
+  }
 });
 
 module.exports = blogRouter;
